Fix 1 sec pipewire delay by using centos 8 alsa-pulse rpm--
"pipewire" automatically installs "pipewire-alsa", but this is removed and
"alsa-plugins-pulseaudio" used instead.  This was deprecated in Centos 9,
but looks like we need it- so install it manually and telling dnf to
ignore the missing "pulseaudio" package.

This solution was proposed in this thread:
https://old.reddit.com/r/davinciresolve/comments/phizng/significant_audio_latency_on_linux/

And it seems to work!  Since the rpm is downloaded without TLS, I use a
pinned sha hash.

Tested with two containers:
RESOLVE_BASE_CONTAINER_IMAGE=quay.io/centos/centos:stream9
RESOLVE_BASE_CONTAINER_IMAGE=almalinux/9-base

Host is also running pipewire, but not sure that matters.