A fix for (Arch/Manjaro) - "magic bytes" in container breaks build.
See Issue #16 for discussion.  Basically this checks to see if the
problematic bytes are in the header, and if so, patches them out.
Confirmed working on Ubuntu 22.04, Pop!_OS, and Manjaro.

Thanks to @hanziness and @cowmix for helping to find the cause of the
problem and for testing.  Also thanks to @theassassin and the AppImage
folks for identifying and fixing, which will hopefully be addressed by
BMD if they use a newer AppImageKit, at which point this check will
fail and this commit won't be needed on those systems.

Still not sure why --endian=big is needed on x86_64 but apparently it is..

¯\_(ツ)_/¯